requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-social-share==2.3.0
modeltranslation==0.25
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
